Platformer Group Project
This was a basic platformer game made over the course of several weeks with a group of other students in my class. I was responsible for gameplay mechanics (such as movement, collectibles, jumping/jumppads, etc.) and I also served as the project manager.

The game was made using Blueprints as it was the most comfortable option for the other students involved. Several concepts were cut due to time, but the basics are there.